chore: refactors theme provider to seperate files
