+++ website/components/ThemeProvider/GlobalStyles.tsx
@@ -1,0 +1,14 @@
+import { createGlobalStyle } from 'styled-components'
+
+const GlobalStyles = createGlobalStyle`
+  * {
+    font-family: 'Nunito Sans', sans-serif;
+  }
+
+  h1 {
+    font-weight: 900;
+    font-style: italic;
+  }
+`
+
+export default GlobalStyles
