+++ website/components/ThemeProvider/index.ts
@@ -1,0 +1,2 @@
+import ThemeProvider from './ThemeProvider'
+export default ThemeProvider
